refactor: modularizes code
